feat: move type & themeColorMap & switchPosition to App.vue
